Display labels on map

+++ app/forms.py
@@ -10,7 +10,7 @@ from django.utils.translation import ugettext_lazy as _
 
 class SearchForm(forms.Form):
     query = forms.CharField(label='Query', max_length=200, required=False)
-    distance = forms.IntegerField(label='Distance', max_value=200, required=False)
+    distance = forms.FloatField(label='Distance', max_value=200, required=False)
     title = forms.CharField(label='Title', max_length=200, widget=forms.TextInput(attrs={"onkeyup" : "searchOpen('title')"}), required=False)
     year = forms.IntegerField(label='Year', max_value=2020, min_value=1880, required=False, error_messages = {'min_value': "Cinematography was only invented in 1880's."})
     production_company = forms.CharField(label='Production Company', max_length=200, widget=forms.TextInput(attrs={"onkeyup" : "searchOpen('production_company')"}), required=False)
